+++ tslib/src/functions/check-marker/check-marker.v1.spec.ts
@@ -6,7 +6,7 @@ import { IAsyncTableService } from '../../azure/storage-helper'
 import { ICheckFormService } from './check-form.service'
 import * as R from 'ramda'
 import { ILogger } from '../../common/logger'
-import { ICheckMarkerFunctionBindings } from './models'
+import { ICheckMarkerFunctionBindings } from './marker-models'
 import checkSchema from '../../schemas/complete-check.v1.json'
 
 const TableServiceMock = jest.fn<IAsyncTableService, any>(() => ({
