+++ tslib/src/functions/check-marker/index.ts
@@ -2,7 +2,7 @@ import { AzureFunction, Context } from '@azure/functions'
 import { MarkCheckMessageV1 } from '../../schemas/models'
 import { performance } from 'perf_hooks'
 import * as V1 from './check-marker.v1'
-import { ICheckMarkerFunctionBindings } from './models'
+import { ICheckMarkerFunctionBindings } from './marker-models'
 const functionName = 'check-marker'
 
 const serviceBusQueueTrigger: AzureFunction = async function (context: Context, markCheckMessage: MarkCheckMessageV1): Promise<void> {
